refactor: Move scope and binding types to `scope.rs`
