+++ crates/ruff/src/rules/pep8_naming/rules/dunder_function_name.rs
@@ -3,8 +3,8 @@ use rustpython_parser::ast::Stmt;
 use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::helpers::identifier_range;
+use ruff_python_ast::scope::{Scope, ScopeKind};
 use ruff_python_ast::source_code::Locator;
-use ruff_python_ast::types::{Scope, ScopeKind};
 
 /// ## What it does
 /// Checks for functions with "dunder" names (that is, names with two
