+++ crates/ruff/src/rules/pep8_naming/rules/invalid_first_argument_name_for_method.rs
@@ -3,7 +3,8 @@ use rustpython_parser::ast::{Arguments, Expr};
 use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::function_type;
-use ruff_python_ast::types::{Range, Scope};
+use ruff_python_ast::scope::Scope;
+use ruff_python_ast::types::Range;
 
 use crate::checkers::ast::Checker;
