+++ crates/ruff/src/rules/pycodestyle/rules/lambda_assignment.rs
@@ -4,8 +4,9 @@ use ruff_diagnostics::{AutofixKind, Availability, Diagnostic, Fix, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::helpers::{match_leading_content, match_trailing_content, unparse_stmt};
 use ruff_python_ast::newlines::StrExt;
+use ruff_python_ast::scope::ScopeKind;
 use ruff_python_ast::source_code::Stylist;
-use ruff_python_ast::types::{Range, ScopeKind};
+use ruff_python_ast::types::Range;
 use ruff_python_ast::whitespace::leading_space;
 
 use crate::checkers::ast::Checker;
