+++ crates/ruff/src/rules/pyflakes/rules/unused_annotation.rs
@@ -1,6 +1,6 @@
 use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, violation};
-use ruff_python_ast::types::ScopeId;
+use ruff_python_ast::scope::ScopeId;
 
 use crate::checkers::ast::Checker;
