+++ crates/ruff/src/rules/pylint/rules/await_outside_async.rs
@@ -2,7 +2,8 @@ use rustpython_parser::ast::Expr;
 
 use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, violation};
-use ruff_python_ast::types::{FunctionDef, Range, ScopeKind};
+use ruff_python_ast::scope::{FunctionDef, ScopeKind};
+use ruff_python_ast::types::Range;
 
 use crate::checkers::ast::Checker;
