+++ crates/ruff_python_ast/src/function_type.rs
@@ -2,7 +2,7 @@ use rustpython_parser::ast::Expr;
 
 use crate::context::Context;
 use crate::helpers::{map_callable, to_call_path};
-use crate::types::{Scope, ScopeKind};
+use crate::scope::{Scope, ScopeKind};
 
 const CLASS_METHODS: [&str; 3] = ["__new__", "__init_subclass__", "__class_getitem__"];
 const METACLASS_BASES: [(&str, &str); 2] = [("", "type"), ("abc", "ABCMeta")];
