+++ crates/ruff_python_ast/src/lib.rs
@@ -9,6 +9,7 @@ pub mod logging;
 pub mod newlines;
 pub mod operations;
 pub mod relocate;
+pub mod scope;
 pub mod source_code;
 pub mod str;
 pub mod types;
